+++ CHANGELOG.md
@@ -142,6 +142,16 @@
   a search job instead of blocking and returning results. 
   The `Service.search(query, inputArgs, outputArgs)` has been removed.
 
+* Method OutputServer.setsslRootCAPPath has been renamed to setSslRootCAPPath.
+
+* Method SavedSearch.getDispatchReduceFreq, which returned a String, has been replaced
+  with getDispatchReduceFrequency, which returns an integer.
+
+* setRestrictToHost has been removed from TCP and UDP inputs.
+
+* Settings.setMgmtHostPort renamed to setMgmtPort, and Settings.getMgmtPort now returns
+  an integer.
+
 ### Bug fixes
 
 * The `Service.versionCompare` method has been fixed to work as expected.